requirements.txt
flask==3.0.0
gunicorn==22.0.0
python-dotenv==1.0.1
paho-mqtt==2.1.0
requests==2.32.3
sqlalchemy==2.0.34
psycopg2-binary==2.9.9